Move outdated code to separate module
